Fix bug in logic for storing OppfolgingstilfelleArbeidstaker
In addition to storing OppfolgingstilfelleArbeidstaker when person is Arbeidstaker at end of latest Oppfolgingstilfelle and OppfolgingstilfelleArbeidstaker is DialogmoteKandidat, also store Oppfolgingstilfellearbeidstaker if the same condition is false for the latest Tilfelle and true for a Tilfelle that is not not latest one. This fix should ensure that a DialogmotekandidatEndring is not generated incorrectly in the cases where a person first meets the requirement for Dialogmotekandidat and a DialogmotekandidatEndring is PLANLAGT, but then there is a new Tilfelle and the person is no longer meets the requirement.